backend/requirements.txt
Django
djangorestframework
requests
drf_yasg
psycopg2-binary
gunicorn